fix : PresentRankingList를 다시 props를 통해 상위 컴포넌트에서 data를 받아 ui를 생성하는 것으로 변경해 해결, useRankingItem에 useSuspenseQuery가 아닌 useQuery로 수정
