+++ src/components/Home/PresentRanking/Cotainer/PresentRankingContainer.tsx
@@ -1,5 +1,5 @@
 import RankingTagContainer from '@src/components/Home/PresentRanking/Cotainer/RankingTagContainer';
-import { Suspense, useState } from 'react';
+import { useState } from 'react';
 import {
   StyledPresenetRankingAddItemBtn,
   StyledPresenetRankingAddItemBtnDiv,
@@ -7,11 +7,10 @@ import {
   StyledPresentRankingContainerTitle,
   StyledPrsentRankingDiv,
 } from '@src/components/Home/PresentRanking/Cotainer/StyledPresentRankingContainer';
-import { ErrorBoundary } from 'react-error-boundary';
 import { useLocation } from 'react-router-dom';
 import { PARAMS } from '@src/assets/params';
 import { useRankingItem } from '../Item/useRankingItem';
-import PresentProductList from '@src/components/Home/PresentRanking/Item/PresentRankingItem';
+import PresentProductList from '../Item/PresentProductList';
 
 const PresentRankingContainer = () => {
   const [isVisible, setisVisible] = useState(false);
@@ -19,7 +18,7 @@ const PresentRankingContainer = () => {
   const params = new URLSearchParams(search);
   const rankType = params.get(PARAMS.rankType);
   const targetType = params.get(PARAMS.targetType);
-  const { data } = useRankingItem({ targetType, rankType });
+  const { data, isError, isLoading } = useRankingItem({ targetType, rankType });
 
   const handelToogle = () => {
     setisVisible((prev) => !prev);
@@ -34,15 +33,13 @@ const PresentRankingContainer = () => {
         <RankingTagContainer></RankingTagContainer>
       </div>
       <StyledPrsentRankingDiv>
-        <ErrorBoundary fallback={<div>에러 발생!</div>}>
-          <Suspense fallback={<div>Loading...</div>}>
-            <PresentProductList
-              data={data}
-              isVisible={isVisible}
-              showRankingNumber={true}
-            ></PresentProductList>
-          </Suspense>
-        </ErrorBoundary>
+        <PresentProductList
+          data={data}
+          isError={isError}
+          isLoading={isLoading}
+          isVisible={isVisible}
+          showRankingNumber={true}
+        ></PresentProductList>
       </StyledPrsentRankingDiv>
       <StyledPresenetRankingAddItemBtnDiv>
         <StyledPresenetRankingAddItemBtn onClick={handelToogle}>
