+++ src/components/Home/PresentRanking/Item/useRankingItem.ts
@@ -1,7 +1,7 @@
-import { getFetch } from '@src/api/getFetch';
+import { getFetch } from '@src/api/getBasicFetch';
 import { BASIC_ENDPOINT } from '@src/assets/endpoints';
 import type { Goods } from '@src/types/Goods';
-import { useSuspenseQuery } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 
 export const useRankingItem = ({
   targetType,
@@ -10,7 +10,7 @@ export const useRankingItem = ({
   targetType: string | null;
   rankType: string | null;
 }) => {
-  const { data, isError } = useSuspenseQuery<Goods>({
+  const { data, isError, isLoading } = useQuery<Goods>({
     queryKey: ['ranking', { targetType, rankType }],
     queryFn: () =>
       getFetch<Goods>(BASIC_ENDPOINT.ranking, {
@@ -18,5 +18,6 @@ export const useRankingItem = ({
         rankType,
       }),
   });
-  return { data, isError };
+  console.log(data);
+  return { data: { data: data?.data ?? [] } as Goods, isError, isLoading };
 };
