+++ src/components/Home/PresentTheme/Item/usePresentThemeFetch.ts
@@ -1,4 +1,4 @@
-import { getFetch } from '@src/api/getFetch';
+import { getFetch } from '@src/api/getBasicFetch';
 import { BASIC_ENDPOINT } from '@src/assets/endpoints';
 import type { Themes } from '@src/components/Home/PresentTheme/Item/ThemeType';
 import { useQuery } from '@tanstack/react-query';
