+++ src/components/ThemesProduct/useThemesProductLabel.ts
@@ -2,7 +2,7 @@
 import { useParams } from 'react-router-dom';
 import { BASIC_ENDPOINT } from '@src/assets/endpoints';
 import { useQuery } from '@tanstack/react-query';
-import { getFetch } from '@src/api/getFetch';
+import { getFetch } from '@src/api/getBasicFetch';
 
 type ThemeLabel = {
   data: {
